Move Json Parser to utils
