+++ training/src/main/scala/com/airbnb/aerosolve/training/utils/JsonParser.scala
@@ -1,7 +1,6 @@
-package com.airbnb.aerosolve.training
+package com.airbnb.aerosolve.training.utils
 
-import com.fasterxml.jackson.databind.DeserializationFeature
-import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.databind.{DeserializationFeature, ObjectMapper}
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
 import com.fasterxml.jackson.module.scala.experimental.ScalaObjectMapper
